requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
tavily-python
python-dotenv
pydantic
httpx
